SOC (Separation of concerns);  - Making the book router to define the routes only and adding the book controller to handle rest of the stuffs;
